Save before trying the game without the controllable hero captain
Hero abilites button UI updated with new assets.

+++ Assets/Scripts/Special/UpgradesShopHero.cs
@@ -254,7 +254,7 @@ public class UpgradesShopHero : MonoBehaviour
 
     void SetPlayerGoldText()
     {
-        transform.Find("Money/MoneyText").GetComponent<TextMeshProUGUI>().text = saveManager.money.ToString();
+        transform.Find("Money/MoneyText").GetComponent<TextMeshProUGUI>().text = ((int)saveManager.money).ToString();
     }
